+++ Sources/SubscriptionExample/Models/Constants.swift
@@ -1,5 +1,12 @@
-// Constants.swift
-// Package constants
+//
+//  File name: Constants.swift
+//  Project name: subscriptionexample
+//  Workspace name: subscriptionexample
+//
+//  Created by: nothing-to-add on 23/05/2025
+//  Using Swift 6.0
+//  Copyright (c) 2023 nothing-to-add
+//
 
 import Foundation