+++ Sources/SubscriptionExample/Models/MockProduct.swift
@@ -1,5 +1,12 @@
-// MockProduct.swift
-// Mock product model for UI display during development
+//
+//  File name: MockProduct.swift
+//  Project name: subscriptionexample
+//  Workspace name: subscriptionexample
+//
+//  Created by: nothing-to-add on 21/05/2025
+//  Using Swift 6.0
+//  Copyright (c) 2023 nothing-to-add
+//
 
 import Foundation
 import StoreKit